requirements.txt
Django==3.2.19
google-generativeai==0.3.2
pymongo==3.12.3
python-dotenv==1.0.0
werkzeug==2.3.7
sqlparse==0.2.4
djongo==1.3.6
